+++ avocado/utils/network/exceptions.py
@@ -1,0 +1,4 @@
+class NWException(Exception):
+    """
+    Base Exception Class for all exceptions
+    """
